Use companyAdminSecurityService for each role check
