+++ src/main/java/com/coherentsolutions/pot/insuranceservice/security/service/CompanyAdminSecurityService.java
@@ -1,0 +1,63 @@
+package com.coherentsolutions.pot.insuranceservice.security.service;
+
+import java.util.UUID;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CompanyAdminSecurityService{
+
+  public boolean canAccessCompanyUsers(UUID companyId) {
+    return canAccessCompany(companyId, "ROLE_FUNC_COMPANY_USER_MANAGER");
+  }
+
+  public boolean canAccessCompanyPlans(UUID companyId) {
+    return canAccessCompany(companyId, "ROLE_FUNC_COMPANY_PLAN_MANAGER");
+  }
+
+  public boolean canAccessCompanyInsurancePackages(UUID companyId) {
+    return canAccessCompany(companyId, "ROLE_FUNC_COMPANY_INSURANCE_PACKAGE_MANAGER");
+  }
+
+  public boolean canAccessCompanyManagement(UUID companyId) {
+    return canAccessCompany(companyId, "ROLE_FUNC_COMPANY_MANAGER");
+  }
+
+  public boolean canAccessCompany(UUID companyId, String requiredRole) {
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+    if (authentication == null) {
+      return false;
+    }
+
+    boolean belongsToCompany = checkIfBelongsToCompany(authentication, companyId);
+
+    if (isAppAdmin(authentication)) {
+      return true;
+    }
+
+    boolean hasRequiredRole = authentication.getAuthorities().stream()
+        .anyMatch(a -> a.getAuthority().equals(requiredRole));
+
+    return hasRequiredRole && belongsToCompany;
+  }
+
+  private boolean isAppAdmin(Authentication authentication) {
+    return authentication.getAuthorities().stream()
+        .anyMatch(a -> a.getAuthority().equals("ROLE_APPLICATION_ADMIN"));
+  }
+
+  private boolean checkIfBelongsToCompany(Authentication authentication, UUID companyId) {
+    if (authentication.getPrincipal() instanceof Jwt) {
+      Jwt jwt = (Jwt) authentication.getPrincipal();
+      String tokenCompanyId = jwt.getClaimAsString("company_id");
+      return tokenCompanyId != null && tokenCompanyId.equals(companyId.toString());
+    }
+    return false;
+  }
+
+
+
+}
